Move "wormhole" transition implementation into "addon" folder
